Move interactors into corresponding packages for organization
